Decode encoded-words in subjects

+++ mta/processing.h
@@ -10,7 +10,7 @@
 void tabsToSpaces(char * const text, const size_t len);
 int getHeaders(unsigned char * const data, size_t * const lenData, struct emailInfo * const email);
 
-void decodeEncodedWord(unsigned char * const data, size_t * const lenData);
+void decodeEncodedWord(unsigned char * const data, uint8_t * const lenData);
 
 void moveHeader(unsigned char * const data, size_t * const lenData, const char * const needle, const size_t lenNeedle, unsigned char * const target, uint8_t * const lenTarget, const size_t limit);
 unsigned char *decodeMp(const unsigned char * const src, size_t *outLen, struct emailInfo * const email, unsigned char * const bound0, const size_t lenBound0);
